requirements/base.txt
django==1.10

# Configuration
django-environ==0.4.0
whitenoise==3.2.1

# DB
psycopg2==2.6.2

# Time zones support
pytz==2016.6.1

# Date/time strings
python-dateutil==2.5.3

# HTTP
requests==2.11.0

# REST
djangorestframework==3.4.4
django-cors-headers==1.1.0
django-filter==0.14.0
djoser==0.5.0
djangorestframework-jsonapi==2.1.0
